revert renaming of sh_audio
