+++ spharpy/classes/audio.py
