+++ tests/test_spherical_harmonics_audio.py
@@ -1,7 +1,7 @@
 from pytest import raises
-from spharpy.classes.sh_audio import _SphericalHarmonicAudio
-from spharpy.classes.sh_audio import SphericalHarmonicTimeData
-from spharpy.classes.sh_audio import SphericalHarmonicFrequencyData
+from spharpy.classes.audio import _SphericalHarmonicAudio
+from spharpy.classes.audio import SphericalHarmonicTimeData
+from spharpy.classes.audio import SphericalHarmonicFrequencyData
 import numpy as np
 
 
@@ -10,7 +10,7 @@ def test_init_sh_time_data():
     times = [1, 2, 3, 4]
     sh_time_data = SphericalHarmonicTimeData(
         data, times,  basis_type='real', normalization='SN3D',
-        channel_convention="acn", condon_shortley=False,
+        channel_convention="ACN", condon_shortley=False,
         comment="")
     assert type(sh_time_data) == SphericalHarmonicTimeData
 
@@ -20,6 +20,6 @@ def test_init_sh_frequency_data():
     frequencies = [1, 2, 3, 4]
     sh_freq_data = SphericalHarmonicFrequencyData(
         data, frequencies, basis_type='real', normalization='SN3D',
-        channel_convention="acn", condon_shortley=False,
+        channel_convention="ACN", condon_shortley=False,
         comment="")
     assert type(sh_freq_data) == SphericalHarmonicFrequencyData
